Exit quicker, if there is nothing to do. (LP: #1287674)
It appears that packages are not using content-hub at the moment,
yet allocating/opening dconf database, processing keys and
iterating across each type of peer takes a significant amount of
time (on slow IO, low clock speed ubuntu emulator). Ideally it
should be improved to be more performant when actually in use,
this patch merely short-circuits any operations when content-hub
is not used at all.